requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Couchbase
couchbase>=4.3.0

# Kafka + Avro
confluent-kafka>=2.4.0
avro>=1.11.0

# Keycloak / JWT
fastapi-keycloak-middleware>=1.2.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
